requirements.txt
python第三方库：  
SQLAlchemy==1.2.7  
PyMySQL==0.9.3  
pandas==0.23.0  
pyyaml==3.12  
pycryptodome==3.8.1  
pytorch==1.0.1  
torchvision==0.2.2  
pillow==5.3.0  
opencv-python==4.1.0.25  
pytesseract==0.2.6  

非python第三方库：
tesseract-orc==4.0.0